+++ database/migrations/2026-07-31-add-apple-registration-source.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+ALTER TYPE registration_source ADD VALUE IF NOT EXISTS 'apple';
+
+COMMIT;